NOTE COMMIT:
In the previous commit, we found translation bug in order to fix it we need to find the correct translation and rotation axis. But we can't trust the cog/com of the ring. In the current method we assumed rotation/translation axis points from frame1 cog/com to frame2 cog/com, this has a drawback. The sign of the rotation/translation may not be correct because, when ring might move backwards, negative direction becomes positive direction.

fix for the translation bug: lets assume some direction in the space and define rotation and translation with respect to that direction. This reference direction should approximately represent direction of the motion but if it does not then we will just get opposite sign. This might work for linear track systems but for circular track systems it might not work. For now we will take this reference vector from some end atoms in the track.

FUTURE WORK:
 * Implement fix for the translation bug